Langchain Beginners/LangchainPrompts/gemini_model.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv

load_dotenv()

llm = ChatGoogleGenerativeAI(
    model="gemini-2.5-flash",
    temperature=0.7,
    max_tokens=None,
    timeout=None,
    max_retries=2,
)

response = llm.invoke("What are the benifits of using Langchain with Gemini Model?")

print("Response from Gemini Model: ", response)
